ingest ops helper funcs
